made it possable to add subjobs or phantom classes
to anyone looking at this github and wants to complain about subjobs being taken off the offical site, this is not offical and for use by LNRW for their tabletop games.
changes made
made it so a class can have level and hd based things turned off effecting:
bab, saves, exp, HD and therefor HP

need to add skills to this? not sure will check later